Move code to plugins directory
